+++ lib/src/data/dialog_info.dart
@@ -1,3 +1,5 @@
+import 'package:flutter_smart_dialog/src/custom/custom_dialog.dart';
+
 import 'base_dialog.dart';
 
 class DialogInfo {
@@ -5,6 +7,7 @@ class DialogInfo {
     required this.dialog,
     required this.backDismiss,
     required this.isUseAnimation,
+    required this.type,
     required this.tag,
   });
 
@@ -13,6 +16,8 @@ class DialogInfo {
   final bool backDismiss;
 
   final bool isUseAnimation;
+
+  final DialogType type;
 
   final String? tag;
 }